backend/requirements.txt
anthropic>=0.52.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
python-dotenv>=1.0.0
pydantic>=2.10.0
twilio>=9.0.0
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
